// frontend/src/core/api/generated/model/shipmentRowType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type ShipmentRowType = (typeof ShipmentRowType)[keyof typeof ShipmentRowType]

export const ShipmentRowType = {
  INBOUND: 'INBOUND',
  TO_SITE: 'TO_SITE',
  TRANSFER: 'TRANSFER',
  OUTBOUND: 'OUTBOUND',
  RETURN: 'RETURN',
} as const